fix: fix  Cast to ObjectId failed

+++ src/resources/tasks/task.db.repository.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const { Task } = require('./task.model');
 const { NOT_FOUND_ERROR } = require('../../errors/appErrors');
 const ENTITY_NAME = 'task';
@@ -7,6 +8,12 @@ const getAllbyBoard = async boardId => Task.find({ boardId });
 const getAll = async () => Task.find({});
 
 const getById = async (boardId, taskId) => {
+  if (
+    !mongoose.Types.ObjectId.isValid(boardId) ||
+    !mongoose.Types.ObjectId.isValid(taskId)
+  ) {
+    throw new NOT_FOUND_ERROR(ENTITY_NAME, { boardId, taskId });
+  }
   const task = await Task.findOne({ _id: taskId, boardId });
   if (!task) {
     throw new NOT_FOUND_ERROR(ENTITY_NAME, { taskId });
@@ -17,6 +24,12 @@ const getById = async (boardId, taskId) => {
 const create = async task => Task.create(task);
 
 const update = async (boardId, taskId, data) => {
+  if (
+    !mongoose.Types.ObjectId.isValid(boardId) ||
+    !mongoose.Types.ObjectId.isValid(taskId)
+  ) {
+    throw new NOT_FOUND_ERROR(ENTITY_NAME, { boardId, taskId });
+  }
   const task = await Task.updateOne({ _id: taskId, boardId }, data);
   if (!task) {
     throw new NOT_FOUND_ERROR(ENTITY_NAME, { taskId, boardId, data });
@@ -25,6 +38,12 @@ const update = async (boardId, taskId, data) => {
 };
 
 const remove = async (boardId, taskId) => {
+  if (
+    !mongoose.Types.ObjectId.isValid(boardId) ||
+    !mongoose.Types.ObjectId.isValid(taskId)
+  ) {
+    throw new NOT_FOUND_ERROR(ENTITY_NAME, { boardId, taskId });
+  }
   const task = await Task.deleteOne({ _id: taskId, boardId });
   if (!task) {
     throw new NOT_FOUND_ERROR(ENTITY_NAME, { taskId, boardId });
